// app/client/library-modules/domain-models/user/ProfileData.ts
export type ProfileData = {
  profileDataId: string;
  userAccountId: string;
  firstName: string;
  lastName: string;
  email: string;
  dob: string;
  occupation: string;
  phone: string;
  emergencyContact: string;
  employer: string;
  workAddress: string;
  workPhone: string;
  profilePicture: string;
  carMake: string;
  carModel: string;
  carYear: string;
  carPlate: string;
};